truly optimal
I refined it a bit more.  now it doesn't double back on itself, which both saves computation time and means I don't haff to worry about paths that should be identical not being identical.  saving it in a fashion that accounts for this new method also saves on disc space.  I'm very happy with it now.