All together now
I made it stitch together the different sections!  it looks kind of bad now.  I need to work on the mesh bilder.  actually, it mite be fine if I don't use any logs in the error function.  but it would still be nice to improve on this a bit.  maybe if I bild it from the south pole outward.