Unemployed cells
I made the mesh generation node based instead of cell based.  it actually works better this way, tho not as better as I would have liked.  now there are only a few inside out cells.  I think it's time to call this as good as it will get and work on the optimization.