advanced tecnekes
I made a few minor but significant changes here.  mostly, I fixd the bug where the parallels were curving the rong way.  I rotated the plot 90° because that apparently fits the final map better.  I added a check to ensure the elements are all rite-side-out when we switch to the strict energy function (they're not, but I think I know how to fix that). I implemented a stricter line search criterion but haven't really tried it out yet. I fixd a bug where I realized np.log doesn't work on Variables. and I *tried* to implement some scaling on the gradients, but I didn't find a skeme that seemed like it was going to work.  so my new plan to speed convergence is to reduce the number of nodes for the first few descents and interpolate them in afterward.