don't crowd the poles
I made the mesh skeleton better, so there are fewer points near the south pole.  it does tecnicly now take longer to converge, but I think it's probably just converging farther.  what's good is that it is no longer at all limited by the jostling of the circumpolar nodes.  instead it is now mostly limited by two seemingly random nodes in southern africa.  I'm not sure about what that is.