We're going all the way
I got it to do the entire fitting rutine including the strict objective function.  I was a fool; the negative stretches were coming from the line search.  I removed the error message associated with that and had it return infinity instead.  now it actually goes all the way to the end without a hitch!  well, it's worth noting that the second phase, where it sticks to the reduced mesh space but switches to the strict objective function, currently terminates in 0 iterations.  that sort of makes sense, since they should be similar when near their minima, but I'm surprised that it's not, like 2 or 3 iterations.  I want to look into that.  it's possible that the initial step size is too short.  anyway, this is looking pretty okey, and I think a good next step is to dive into the convergence properties to make sure it's going as fast as it can before I go back to hier resolutions.