look at it go!
I made a little status tracker thing to help me see how the convergence is going.  it seems to be working decently well, tho there are still some issues.  for one thing, it seems to be attracted toward a map that is about 2 times smaller than actual size, which is weerd.  the minimization pass also always finishes in 0 iterations, which is super weerd.  it's also definitely converging longer than is necessary.  I want to try a gradient-magnitude-based termination condition.  and also include some form of regularization.
 and also change the colorscales