It looks flat
I changed it to a gradient-magnitude-based stopping condition.  I dislike the fact that I need to know the order of magnitude of the gradient ahead of time (I couldn't find a good way to define it relatively), but it is certainly much more logical than the improvement-based one, and it now stops at about the rite time.