twice is enuff
it turns out that after a single pass at optimization, the lenient and strict cost functions are essentially identical.  that's relieving, since that was my intention, but I must say that I'm surprised it's true to such an exact degree.  anyway, since the twoth optimization pass doesn't do anything, I can simplify the code a bit.