the best I can do without internet
I set it up to save and load the various weight arrays as TIF files.  now the map projections it generates are actually properly weitet!  not that I can really tell whether it's working with this low resolution.  I tried turning it up and hit a memory error.  It seems I'll haff to go to sparse arrays after all.