proper downsampling!
I fixd the downsampling.  I also moved the function into create_map_projection.py because I rememberd that there's a very similar function in find_drainage_divides.py and they're not quite similar enuff for me to make a single general utility function for both but I don't like having a util function that sounds relevant to find_drainage_divides and not using it.  so in create_map_projeciton in goes.