sparsely populated
I implemented sparse arrays.  they are extremely slow when implemented in python, as it turns out.  it looks like to acheve my dream, I'll need to turn to ctypes after all (well, or rite it in Java, but I think it's a little late for that).  oh well.  should be an interesting learning experience.