cytonic processing
I implemented the sparse arrays in C instead of Python.  now it can dip into the shadow realm to speed up communications.  it's insanely fast now compared to the Python version and even compared to the original dense numpy version!  this is very exciting.  and now I know how to use ctypes!  I'm so excited.