pyramid scheme!
I did the same thing as last time, but actually good this time.  it significantly speeds up the first stage of the optimization, and yeelds visibly better results.  it did scope-balloon a little bit, as I ended up implementing matrix multiplication in my sparse arrays so that the reduction operation could be represented as a single matrix rather than a weerd function, but I think this will serve me well in the long run, since now I can easily extend this sparse array usage to my constraints later on.