vectorized skeleton now instantaneus
yeah, it's a little bit absurd how much of a speedup you get from doing the same thing in C.  well, I did also change the algorithm somewhat.  instead of finding duplicates by checking every previus item in the list, it's now sorting the list first, which shaves off a factor of O(n/logn), where n is ~10^3.  so that's not noting either!  anyway the point is, it works great now!