interruptions now block information
I made it so that the weight maps are more complicated such that you will never agen have a fuzzy edge of perceivd importance at the edge of the mesh even tho there's not really any land there.  because land only affects cells in the same section now.  it's a little weerd around the glue borders, but I think it's fine.