Plaintext maps
I spent some time fleshing out the way the maps are saved.  main things are that I added a plaintext format for those who don't know how to read HDF5, and I added the projected border.  this required me to rite the border projection code, which will also be useful for adding constraints, but I have yet to get that far.  in addition, I figured out how to label HDF5 axes and collums, so now the position tables are a bit more readable.

there are some problems.  most notably, the stop condition seems to be triggering early now.  and also the saving part takes a very long time.  and also trying to read the output throws a null pointer dereferencing error.  but I really want to push something, so here you go.