the sharp outline
I workd thru problems with the border not displaying properly to make it so it does.  I'm quite happy with it; it looks good.  it turns out that there were kind of a lot of problems leading to extra lines all over the map, including some real doozies (like h5py defaulting to float32 instead of float64).  but it's all g-munny now.