more aggressive fitting
I've tuned the bounded map fitting by making it eliminate inside-out cells earlier on and having a better stop condition on the polytope projection.  it runs smoothly and without errors now, but it seems to be terminating too early, and also looks like it may be getting stuck in a local minimum.