ci1ji1 ting2 che1!
I refined the stop condition for the last time.  this is a new one that calculates the fraction of the step magnitude that gets removed by projection.  since there's so many dofs, it never goes below .05 or so, but a relatively coarse tolerance on it seems to terminate the algorithm at about the rite time.