that's odd.
I made the mesh odd instead of even so that cuts along the prime meridian fall in the middle of a cell instead of along the edge of one.  I also redid many of the imports to use NDArray instead of np.ndarray and to use unqualified names for math functions.