eigenspace
I changed it so the quadratic minimization transforms into a better coordinate space and then performs ordinary projection.  It should require fewer iterations when the eigenvalues of the hessian have a lot of variation, since it makes up for the fact that the FDBPG routine doesn't use any twoth-order information.  and it doesn't cost anything extra since I need the eigendecomposition regardless (for when I compute (A + λI)^-1).  it's still kind of slow... but this seems like it may be all I can do... I'll tweak it some and then try running it overnite.