logging
I implemented a more robust logging system since the print statements were bothering me.

I think I'm coming to understand why exactly this is hard.  at the end of the day, I'm using the dual problem because that converts the valid area to the positive orthant, but the hessian of the dual problem is not and never will be isotropic.  it is in fact heavily skewed, which is why gradient descent is taking hundreds of thousands of iterations.  so converting to eigenspace is pointless because when I convert to the dual problem I essentially set it right back.  Instead, I need a general twoth-order technique that can handle a positive orthant constraint.  I'll continue shopping around online; there do appear to be alternatives to the FDBPG method.