the puny stepsize was a safety, apparently
I implemented a variable line search for the fast dual-based proximal gradient algorithm.  it does make it converge faster in general, but the line searches don't converge anymore.  I'm gessing it's because I needed to fudge the indicator function conjugate (I removed the domain limitation) because the fista step extrapolation thing plus good step sizes was causing it to swing into the illegal region.  I suppose now it's going into the illegal region anyway, tho.  I could remove the fista, but then I would have to make my stop condition better, and I'm still not sure how to do that since I'm not computing the gradient of the dual objective function (and it's not even differentiable, really).  so I think I'd rather give up on FDBPG and pivot to ADMM, which I feel should be faster, and which has better documentation that I can find.

I also included the golden ratio in my code because it's the best number ever