finally bounded optimization that works!
I ditched ADMM because I couldn’t figure out what was rong with it.  maybe ADMM is just inherently terrible for this problem.  I looked into interior point methods, and it turns out they’re dirt simple.  I implemented it, and it seems to be working *super* well.  the key here is that the additional level of iteration is now at an upper level, not a lower one (I call the main minimize function several times with different "barrier parameters"), so I don’t have to worry about making it ultra fast, and also