it is finished
I polished off the water map and also fleshd out the political map and the terrestrial ecoregions map.  it required a bit of extra code to allow for multicolored shapefiles, but I’m really happy with how it turnd out.  I think the ecoregions map in particular looks pretty sweet.