much much nicer drainage divides
I made it decimate drainage divides instead of doing the exact simplification I had written previusly.  this reduces the number of vertices by a *lot* and also makes it look less annoyingly gridular.

I also changed the data reduction operation for when I downsample the heightmap from mean to median, which seems to get better results (in particular it stops trying and failing to draw the line thru tiny islands), forced all segments over ocean to be perfectly strait, and allowed it to cross the Panama and Suez Canals (I didn’t even realize this was an issue until just now).