localization!
I made it output projection files in four different languages: english, spanish, japanese, and pandunia.  is this necessary?  well, no.  is this useful?  probably not.  but it makes me happy.

I also added a PNG of the basic map that gets shown while it’s optimizing, so that’ll be good to put on the README.  unfortunately, I can’t actually make maps rite now because it turns out I broke the border by decimating it.  so I have to fix that.

it also turned out that the border of the map was getting plotted thrice, so got rid of that.