fixed decimation
the decimation was resulting in a path in which every vertex appeared twice.  I’m frankly embarassed I didn’t notice that sooner.  so I fixd that and also made sure the points at ±180° in longitude are immovable.

I also made the inverse projection faster while I was at it.