a little to the left
I adjusted the ways it handles bounds, because it was returning slitely uncentered projections, which was just annoying.  I also refactord some things slitely so that they use the Mesh class more often.  I also went ahead and doubled the fineness of the bounds, since it apparently has enuff error to bother me.