committed to excellent service
I made some minor adjustments to the outputs to make my Java implementation (and thereby other peeple's implementatiosn) somewhat easier.  specificly, I made it so the inverse raster always gives you a point *in* a valid cell (never on the edge of a section's defined region), and I made it so the gradients at shared nodes are calculated in a simpler and twoth-order accurate way, and I made it so there are never any nodes in the mesh whose gradient can't be calculated somehow.