layerd inverse rasters
I made each section have its own inverse raster.  it's a shame because this makes the inverse implementation somewhat more complicated, but it seems to be necessary for the raster to work around the poles, where a point can easily be interpolated onto an input that's not enclosed by the mesh of the section to which it truly belongs.