leave the north pole alone
I set the countries map to cut near the north pole but not all the way to it, because I fear the way it is is too complicated; in introduces some unnecessary exceptions to my Java code.  the map is slitely worse, but the implementation will be much better I think.