allow non-pole-oriented meshes
I realized that the Elastic Earth III projection was suffering from the fact that one of its tears ends on a pole.  it creates a lot of exceptions that I don't really want to withdeal.  I considerd having it stop short of the pole, but I think it's actually better for the map projection to go *thru* the pole.  but that requires having the "glue tripoint" be not on a pole, which was not previusly supported.  so that's what all this is: adding support for tears that go thru the pole basicly.  part of that also involves adding more nodes around the pole, which I think can only help.