less duplication around straits
I reduced the size of the square around each major strait that gets duplicated about the border.  this was motivated by a desire to sever the bering strait square from the pole trimming, but also in general I think 1800km was too much now that I'm really looking at it.