un-object-orient mesh building
I've come to jump on the parade float against object-oriented programming.  it doesn't make any sense to categorize functions as top-level or methods of Section here; especially given how many are static.  it's arbitrary and makes the code structure harder to understand.  so I'm taking all of the methods away from Section except its initializer.  naturally it still makes sense to have Section as a data container. but as just a struct, not an object.