no more extraneus cuts
previusly, I had it set so that the cell containing the end of the interruption for a given projection would always be shared between the two adjacent sections.  this would inevitably catch some of the cut itself, meaning that when the cuts were plotted, there would be an extra unnecessary line that stuck out into the uncut region.  I want to get rid of that.  so I'm adding an automatic adjustment to the cuts that snips off that extra part, such that the interruptions will always end right at the edge of a cell.  part of that involved adjusting the tangency rules, which was fairly challenging.  it's still not exactly rite, but I think this is a god start.  I am god.