use degrees in build_mesh.py
previusly, build_mesh.py was the only script that used radians by default; everything else was in degrees.  this was because it has the stereographic projection part, which is much easier in radians.  but I was afraid the conversions were causing roundoff issues, so I switched it over, with the exception of the specific functions that relate to the stereographic projection.  that wasn't what was causing the roundoff, as it turns out.  or at least, that wasn't the only thing causing roundoff issues.  but whatever.  I think it's simpler like this, so I'm leaving it.