hackily fix a very minor bug
there was one instance in australia where the cut touchd a cell border, bounced back the other way, then went into that cell to die.  the algorithm was taking the second entry as the end of the path even tho the point of tangency would have made more sense.  I fixd it by changing the binning tie breaking direction.  of course, this fixes it for the case of tangency on the north of the final cell, but now won't work on the south of the final cell.  but that's fine; it's a rare phenomenon so there's no need to shinpai overmuch.