no more radians!
I changed all of the angles in create_map_projection to be degrees instead of angles in the hopes of fixing roundoff.  it got most of it, but I also needed to manually round a couple of linspaces in build_mesh to get all the numbers to look nice and pretty.