fix crash
this was an annoyingly hard bug to find.  but I found it.  a shame I didn't notice it looking at the build_mesh plots, but imd it was a pretty easy oversight.  hm, why is oversight the noun form of overlook?  it's also the noun form of oversee.  that seems nonideal...  but I gess overlook as a noun is a completely separate thing.