minor initial condition simplification
I made the code to build the mesh not do a weited comparison of the glue border and cut border, but just treat the whole border of each section at once.  it's a bit simpler, and only minorly impacts the initial condition.  the effect on the final product is imperceptible, and so small on III specificly that the PNG literally doesn't have a bit out of place.  it takes a little bit longer to converge, but it's really not a big deal.