now all it needs is an antenna
I realized that it might *potentially* be helpful for sections where the border goes around the edge to have cells out there so that the edge can be fully projected without modification.  I'm really not sure if this will actually ever be useful, but I don't think there's a reason not to do it, so I've added those cells, and now 1 and 2 look like they have antennas.