stitch up africa
I'm trying out a new configuration of 2 where the interruption doesn't go into Africa.  this increases the distortion but brings the two sides of the Suez together, and makes two continents viewers are likely to recognize rather than just one.  I haven't decided if I'm going to keep it, but I'm leaning no tbh.  the distortion in the Amazon basin and mesopotamia just bothers me.  and the other one is arguably better *because* it's less recognizable.