change section order
the section orders have until now been completely arbitrary.  now they're only half arbitrary: I've set them up such that if any section crosses the antimeridian without containing a pole (sunawachi, it has an inside-out boundary), it will always be last.  that way, you can write an implementation that checks containment while counting all points on the boundary as in, and don't need to worry about points on the antimeridian or poles getting counted just because they're on the inside-out part of the path that technically should contain zero points.  because that one will be last, so any points on the boundary will have been counted as inside a different section already!  I'm not really sure why this affects the optimized map projections at all... I suppose the ordering of the nodes changes which maybe affects roundoff in the matrix inversions.