increase III's aspect ratio
I realized that III lookd really awkward with its small aspect ratio.  it's super rare for a map projection to be that square – even Gall Peters is longer than what I had – and that was exacerbated by the fact that the left and rite sides of the map are quite oceanic despite the selective compression.  more importantly, I realized a lot of countries were being noticeably squeezed east-west.  the old aspect ratio of sqrt(2) was chosen to make my map fit on an A4 peece of paper, but who the heck has ever printed out a world map on an A4 peece of paper?  so I added a bit of horizontal space and clampd down verticly.  the result looks much better in my opinion.  north is more up, antarctica is smaller, and everything is just a little bit less distorted.  as it turns out, golden rectangles are good-looking!