final draft
the Elastic projections are done.  I've re-run all of the scripts, so this now represents the final version of all of the files in this repo.  the only thing left is that I think I'm going to change the name.  I'm currently stuck between "Elastic projections" and "Elasterra projections".  but I can figure that out later; for now what matters is – the ultimate map projection is done!