dilate mesh before saving
in an attempt to smooth out the wrinkles in the mesh and stop it from overlapping itself, I implemented mesh "dilation", where a bunch of nodes are added on the periphery of the already defined region to make interpolation later more continuus.  it makes the full mesh look kind of ugly, but it does *almost* entirely solve the self-intersection issue.  it's still there on the poles, but it's small enuff now that I'm willing to ignore it.