focus distortion more heavily
I found out that I can get a significantly better outcome for Elastic III by decreasing the minimum weit from .03 to .01.  this lowers the material strength of the oceans and allows even more absurd stretching of them to put the landmasses where they belong.  this in turn allows me to further reduce the map size without distorting the continents much.  there are some minor issues I take with the new map – its outline has a little divot next to Alaska that I don't like, and the aspect ratio is larger – 10:6 where it was 13:8.  but the increased size and decreased distortion of the landmasses makes it well worth it.