fine-al map projections
I figured out that I can get even better oceanic compression in III by making the mesh just a smidge finer and then shrinking the bounds to 24×14 Mm².  I don't want to go any farther than this because it already takes quite a while to optimize (and also the problem is getting less and less well behaved as I push to these hier compressions), but I'm happy to take these gains and leave.  I'm a little bit concernd about the cape of good hope, the indus delta, and the gulf of alaska which are all suffering from somewhat noticeable levels of compression here, but the increased north-up-ness of australia, eastern asia, and north america (at the expense of the southern cone but meh) as well as the shrinking of antarctica are both big wins in my book.

it is worth noting that I've noticed some odd behavior here the shape of the map changes nonmonotonicly with the size of the bounds.  like the overlapping region that sometimes turns up between Alaska and Hawaiʻi appears for a 14 500 km high rectangle but not a 14 000 km high one or a 15 000 km high one.  this makes me a bit worried that my algorithm doesn't reliably converge to the global minimum when the bounds are this tite.  but I also really really don't want to dig into that rite now.  so for now I'm going to take my pretty map and leave.